// frontend/src/types/company.ts
export interface CompanySearch {
   currency: string;
   exchangeShortName: string;
   name: string;
   stockExchange: string;
   symbol: string;
}